Moved dtypes
